locli-latex: LaTeX output impl. in Haskell
For now, the EDE library is used to generate LaTeX based on some
template scripts in its own scripting language and some data being fed
to it from within Haskell, with the assistance of some elisp scripts to
drive gnuplot and xelatex. This new feature uses some already partially-
implemented Table (renamed from Org) infrastructure within locli to
render directly in LaTeX without the need for the EDE library.

The implementation of chart generation is being deferred to a second
stage of implementation. As the charts aren't frequently used, they're
being deferred to a follow-up effort. For complete parity with the elisp
scripts and EDE library, this will need a follow-up for chart generation.

This precise commit/etc. represents a rebase squashing a number of
tentative commits in the development process into a single one,
although there is some polishing of the final product as it's viewed
irrespective of commit history) still needed before merging. This is
furthermore expected to be split along more appropriate lines than the
steps taken during development in subsequent rebasings and commit
history cleanups. If this ends up being too awkward, the entire branch
might get thrown away in favor of a different commit history cleanup.