start-leadership: trace drep count and map size
Update TraceStartLeadershipCheckPlus with fields to track the DRep
count and DRep map size. Update LedgerQueries with methods to extract
the information to fill TraceStartLeadershipCheckPlus with. Update
forgeTracerTransform to go about filling in all of the fields of
TraceStartLeadershipCheckPlus with all the results from the accessors.
DRep counts and map sizes get logged in the regularly-scheduled
periodic traces in the final result.